ASoC: msm: qdsp6v2: Avoid sending calibration for DS2 custom topology
Calibration data for DS2 custom topology will be sent when dap
driver is initialized and in bypassed state. When calibration
data is sent from adm driver QTI modules are turned on along
with DAP. This is not expected as QTI modules and DAP module
for DS2 is mutually exclusive.

CRs-Fixed: 676891
Change-Id: Ic9a461ae1418dfcf3b2da491b9917acd08905ca0
Signed-off-by: Krishnankutty Kolathappilly <kkolat@codeaurora.org>